refactor: moved common entites in the core folder, plus added another property to EvaluationResults
