+++ lib/evaluation/core/index.ts
@@ -1,2 +1,4 @@
+export * from './evaluationResults';
+export * from './handlerCheckResult';
 export * from './selectResults';
 export * from './submissionRequirementMatch';
