+++ lib/evaluation/handlers/abstractEvaluationHandler.ts
@@ -3,8 +3,8 @@ import jp from 'jsonpath';
 
 import { Status } from '../../ConstraintUtils';
 import { IInternalPresentationDefinition, WrappedVerifiableCredential } from '../../types/Internal.types';
+import { HandlerCheckResult } from '../core';
 import { EvaluationClient } from '../evaluationClient';
-import { HandlerCheckResult } from '../handlerCheckResult';
 
 import { EvaluationHandler } from './evaluationHandler';
