+++ lib/evaluation/handlers/markForSubmissionEvaluationHandler.ts
@@ -3,8 +3,8 @@ import jp from 'jsonpath';
 import { Status } from '../../ConstraintUtils';
 import { IInternalPresentationDefinition, WrappedVerifiableCredential } from '../../types/Internal.types';
 import PEMessages from '../../types/Messages';
+import { HandlerCheckResult } from '../core';
 import { EvaluationClient } from '../evaluationClient';
-import { HandlerCheckResult } from '../handlerCheckResult';
 
 import { AbstractEvaluationHandler } from './abstractEvaluationHandler';
