+++ lib/evaluation/handlers/sameSubjectEvaluationHandler.ts
@@ -3,8 +3,8 @@ import jp, { PathComponent } from 'jsonpath';
 
 import { Status } from '../../ConstraintUtils';
 import { IInternalPresentationDefinition, WrappedVerifiableCredential } from '../../types/Internal.types';
+import { HandlerCheckResult } from '../core';
 import { EvaluationClient } from '../evaluationClient';
-import { HandlerCheckResult } from '../handlerCheckResult';
 
 import { AbstractEvaluationHandler } from './abstractEvaluationHandler';
